BL_INTERNAL: media: platform: mtk-isp: seninf: Reorganize hardware configuration functions
The SENINF is made of multiple blocks, with their configuration split in
different functions, but without much clarity. Improve this by sorting
functions according to the order of the hardware blocks in the pipeline,
and renaming the functions related to the inputs with a common prefix.
The inputs are named SENINF as the top-level IP core, a
mtk_seninf_seninf_ prefix would be confusing, use mtk_seninf_input_
instead to match the mtk_seninf_input structure.

While at it, drop the unneeded priv pointer passed to the input-related
functions.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Pierre Moos <pmoos@baylibre.com>