BL_INTERNAL: media: platform: mtk-isp: seninf: Configure CSI-2 PHY at runtime PM resume
The CSI-2 PHY configuration only depends on the sources connected to the
CSI-2 inputs and doesn't vary at runtime. It should be performed
globally when starting the SENINF, and not per SENINF input, MUX or
output.

When the driver will support multiple concurrent streams, the
.s_stream() operation will be called once per stream. To prepare for
this, move the CSI-2 PHY global configuration to the runtime PM resume
handler.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Pierre Moos <pmoos@baylibre.com>