BL_INTERNAL: media: platform: mtk-isp: seninf: Store routing in subdev state
Now that we have a subdev state structure that can store routing
information, make use of it to store the active routes. We can drop
the source_pad field in the mtk_seninf_input structure, and use the
helpers to implement .set_routing() and .has_route().

Routing is now initialized in the .init_cfg() operation, which also
initializes formats.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Pierre Moos <pmoos@baylibre.com>