BL_INTERNAL: media: platform: mtk-isp: seninf: Introduce mtk_seninf_mux structure
The SENINF contains separate input blocks (called SENINF) and output
blocks (called MUX). The MUX registers are currently accessed through
the mtk_seninf_input structure that models the inputs, which works for
now as we use the first MUX only.

To prepare for multi-channel support, introduce a mtk_seninf_mux
structure to model the MUX, and plumb it through the driver.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Pierre Moos <pmoos@baylibre.com>