BL_INTERNAL: media: platform: mtk-isp: seninf: Store active formats in subdev state
Now that we have a subdev state structure that can store the format for
each subdev pad, make use of it to store the active formats. We can drop
the custom format fields in the mtk_seninf_input and mtk_seninf
structure, and use the v4l2_subdev_get_fmt() helper to implement
.get_fmt().

The active state is still initialized through the .init_cfg() operation,
but instead of calling it manually, we now need to call
v4l2_subdev_init_finalize() to allocate the active state.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Pierre Moos <pmoos@baylibre.com>