BL_INTERNAL:drm/mediatek: dsi: add ability to select clk_mode during poweron in dts
Normally the DSI clk should be started in low-speed during power on
and then switched to high-speed. We changed this behavior in order
to support the RPi displays. Unfortunately this is breaking the
Startek kd070fhfid015 when it is configured in video mode (only
command mode is working). To fix that let's allow to define the
DSI clk mode in the device-tree.

To be sure we are not breaking anything, the default value will
stay to start the clk as HS.

Signed-off-by: Fabien Parent <fparent@baylibre.com>