drm/mediatek: fix resolution issue for LVDS panel
This configures two DPI registers to fix a resolution issue seen on
mt8365 on an LVDS panel. The issue was that the frame width had twice the
size of the physical screen.

Setting both EDGE_SEL field in DPI_OUTPUT_SETTING, and DPI_DUMMY fixes the
issue. In fact, by comparing the registers from a working image, we found
out that these two registers changes were missing. However, There is still
no clear explanation of why this fixes the resolution issue.

Signed-off-by: Amjad Ouled-Ameur <aouledameur@baylibre.com>