drm/mediatek: mtk_dpi: Change panel preparation timing
From experiments it was found that lcd initialization was happening
too early when compared to the moment LVDS signals are driven. This
could cause some LCD glitches on power-up.

Change panel preparation flow so that it will be executed on good
timing.

Note: experiments were run on a i350_manefaces with a avd-tt70ws
panel. This change should not be ported to other setups without
further consideration.

Signed-off by: Vitor Sato Eschholz <vsatoes@baylibre.com>