dmaengine: mediatek: Add uart irq handler for DMA-mode
The default serial 8250 IRQ handler assumes that the RX buffer is
linear, which is not the case when working in DMA-mode. The DMA RX
buffer is used in a circular fashion, therefore, this commit creates
a new handler to replace serial8250_default_handle_irq() when DMA
is used.

In addition, the interrupt handler needs the knowledge that the DMA
RX engine got reset and hence also reset the corresponding UART buffer
read index which copies out of the DMA buffer.

Signed-off-by: Felix Freimann <felix.freimann@mediatek.com>
Signed-off-by: Vitor Sato Eschholz <vsatoes@baylibre.com>