BL_INTERNAL: gpu: drm: mtk_dpi: set lvdspll rate according to pixel clock
LVDSPLL is the reference clock of VPLL and has its typical value as
300 MHZ. LVDS and DPI pixel clock are derived from VPLL. According to
the LVDS pixel clock frequency, the frequency of LVDSPLL might be
adjusted.

LVDSPLL is initialized by the clock driver while VPLL is initialized
by the lvds driver. To avoid VPLL operating with two different clock
references, moves the LVDSPLL rate setting to the DPI config, instead
of doing it on the LVDS encoder enablement.

Before, the sequence was:
1- LVDSPLL clock initialized at 300 MHZ by clk-mt8365.c
2- VPLL is enabled when LVDS DDP component is configured
3- LVDSPLL clock is updated due to pixel clock at LVDS bridge
   enablement

Now, the sequence is:
1- LVDSPLL clock initialized at 300 MHZ by clk-mt8365.c
2- LVDSPLL clock is updated due to pixel clock when DPI DPP component
   is configured
3- VPLL is enabled when LVDS DDP component is configured

Signed-off-by: Vitor Sato Eschholz <vsatoes@baylibre.com>