[WORKAROUND] thermal: ensure cooling device has correct state
From BL:
The initial curent state of a cooling device is always zero. It will
only be initialized when thermal governor starts to act. If it happens
that the actual current state of the device is also the one targeted by
the governor, it will not be updated an will continue to be zero.

Proper fix is to initialize the curent state of the cooling device before
thermal governor tries to change it. For now, as a workaround, ensure
that the state of the cooling device will be updated even if it matches
the target one.

Signed-off-by: Vitor Sato Eschholz <vsatoes@baylibre.com>