arch: arm64: dts: add labels to mt8365 cooling maps (#8)
* arch: arm64: dts: add labels to mt8365 cooling maps

From BL:
Add labels to the mt8365 cooling maps so that board specific
devicetrees might be able to reference them.

Signed-off-by: Vitor Sato Eschholz <vsatoes@baylibre.com>

* [WORKAROUND] thermal: ensure cooling device has correct state

From BL:
The initial curent state of a cooling device is always zero. It will
only be initialized when thermal governor starts to act. If it happens
that the actual current state of the device is also the one targeted by
the governor, it will not be updated an will continue to be zero.

Proper fix is to initialize the curent state of the cooling device before
thermal governor tries to change it. For now, as a workaround, ensure
that the state of the cooling device will be updated even if it matches
the target one.

Signed-off-by: Vitor Sato Eschholz <vsatoes@baylibre.com>

---------

Signed-off-by: Vitor Sato Eschholz <vsatoes@baylibre.com>